+++ Front-End/src/app/paginas/relatorios/relatorios.component.html
@@ -32,7 +32,7 @@
     </form>
 </div>
 
-<div class="content-20 bg-light pt-4">
+<div class="content-20 bg-light pt-4 bg">
     <div class="row d-flex align-items-center mb-4">
         <div class="col-md-12 d-flex justify-content-around">
             <div class="card-dash">
@@ -69,7 +69,7 @@
                     <mat-icon class="mr-2">paid</mat-icon>
                     <h1 class="h1-title m-0">Transações</h1>
                 </div>
-                <button type="button" class="btn btn-excel d-flex align-items-center">
+                <button (click)="exportExcel()" type="button" class="btn btn-excel d-flex align-items-center">
                     <mat-icon>file_download</mat-icon>
                     Excel
                 </button>
